Modified 'itemDetails.jsx' -stylesheet- - Move from router folder to styles folder. - Name change to match the standard. - Modified the elements to match the proper structure and design.
